requirements.txt
dbus-python
paramiko
ruamel.yaml
